[DebOps] Add new "ldap_password" lookup plugin
On my devel machine, I have a couple of hundred files under
ansible/secret/ldap/credentials/, with no easy way to determine
which are still relevant and which aren't.

This patch adds a new lookup plugin named "ldap_password", which
works in much the same way as the traditional "password" plugin, but
adapted to match the LDAP password storage scheme used by DebOps.

Instead of a single argument (path and options as one string), the new
lookup takes three arguments: base directory, DN, and password options.

The plugin will automatically log the UUID <-> DN mapping in a file in
the base directory as part of the password generation process.