[dovecot] Properly configure LDAP iterate parameters
The problem in a nutshell:

    $ ldapsearch -b dc=example,dc=com -LLL -Q "(uid=foobar)" uid
    dn: uid=foobar,ou=People,dc=example,dc=com
    uid: foobar
    $ ldapsearch -b dc=example,dc=com -LLL -Q "(uid=foobar)" objectClass | grep -q "objectClass: mailRecipient"
    $ echo $?
    1
    $ doveadm user "*" | grep foobar
    foobar
    $ doveadm user "foobar"

    userdb lookup: user foobar doesn't exist

In other words, users without "objectClass: mailRecipient" are included when
requesting all users, but they are *not* valid Dovecot users.

Note that instead of setting "iterate_attrs = uid=user" (which would make
doveadm user "*" return a list of uid's), iterate_attrs is set to
mailAddress, for reasons which should become apparent with later patches.

This causes a change in the output from doveadm user "*", before:

    $ doveadm user "*"
    foobar
    foo
    bar

after:

    $ doveadm user "*"
    foo@example.com
    bar@example.com

But the latter is *anyway* what postldap is looking for, see
"/etc/postfix/ldap_smtpd_sender_login_maps.cf", which includes this:

    query_filter = (&(|
                      (& (objectClass=mailRecipient)(|(mailAddress=%s)(mailAlternateAddress=%s)))
                      (& (objectClass=account) (uid=%u) (host=%d) ))
                     (| (authorizedService=all)
                        (authorizedService=mail:send))
                   )
    result_attribute =
    special_result_attribute = member
    leaf_result_attribute = mailAddress

Note, ldap_smtpd_sender_login_maps.cf is supposed to provide a list
of SASL login names that "own" a given email address (%s). The reason that
the current lookup anyway works is that DebOps Dovecot role has some
funky settings, which will be the topic of the next patch...