[dovecot] Add support for compressed storage
Note that the support is disabled by default. Also: the value of zlib_save
is intentionally hardcoded to "zstd" (even though it isn't supported on
Debian releases older than "stable") so that new installs aren't forced
to use old compression standards (and selecting the compression standard
dynamically depending on the Dovecot version means that the compression
standard would change on dist-upgrade, which is too much of a footgun).