[cron] Update facts script
Update the facts script so that it stores the chosen cron__daemon (similarly
to debops.ntp), which allows us to change the default in the future without
changing the daemon for existing installs.

Note that the somewhat odd definition of "cron__daemon" is intentional (the
second "cron", on the last line, is the actual default when no cron daemon
has been configured, the first invocation just catches old facts.d scripts
which have no "daemon" configured).