[grub] Remove grub__fact_configuration
I consider the use of grub__fact_configuration to be a bit of a footgun.

Right now, the etc/default/grub.d/ansible.cfg.j2 template has the
{{ ansible_managed }} stanza at top, which means that the generated
file will typically read:

  This file is managed remotely, all changes will be lost

The thing is, they won't. If I edit /etc/default/grub.d/ansible.cfg on
a host and, say, add "foobar" to GRUB_CMDLINE_LINUX_DEFAULT, then
that change will be persisted across role invocations, contrary to
what the file itself says.

That's contrary to the notion that the inventory should be the single
source of truth.

Furthermore, the comments for the variable are misleading since they
say that it will ensure that "original parameters [are preserved] on
first run of the role" (they won't be, the role doesn't care at all
what's in the "original parameters").