[icinga_web] Rework and update module handling
Update package versions and implement module removal, update
some schema locations.

Two of the library modules (ipl and vendor) are proper packages
in Bookworm and are hard dependencies for the icingaweb2 package.

On top of that, they will be loaded in preference to modules loaded
via the git mechanism, and the packages modules should be preferred
(DEP-3).

Finally, various modules from git repos can have varying (and
incompatible) requirements as to which other modules should be
installed....so, to make this a bit saner, I've changed the role
to default to using packaged modules.

Unfortunately, the exact packages available varies a lot between
distros, so the module installation is a bit quirky (checking
for the existence of a package and then branching off in one of
two possible directions).