[rabbitmq_server] TLS modernisation
Change the default TLS protocol versions from tlv1.2 + tlsv1.1 to
tlsv1.2 + tlsv1.3 (this is also the upstream default).

Curiously, the rabbitmq_server__ssl_versions variable did not seem to be
in actual use in the current code, so fix that at the same time.

The rabbitmq_server__ssl_ciphers variable was left untouched, but it should
possibly be updated since it relies on running some scripts on the host
and (in a manner which is not evident from the documentation) filtering
out some cryptos.